Extract title.tsx spending cap logic into custom hook
